first pass at modifications asked by lerno
edited the manifest file and provided a compiled box2d library for windows-x64 I use, but imo the best is for others to compile box2d themselves for their own CPU architecture...
Unfortunately cannot provide macos and linux ones at this time.
Changed the raygui icon enum to be a inline ZString so you can pass them directly to raygui functions as zero terminated strings for the "#id#" feature.